S2c: atmosphere(pressure)+wind -> Q16.16 + COLLAPSE every S2 float bridge (the closer)
The FINAL sub-step of the S2 atmosphere group and the CLOSER of the centerpiece:
atmosphere + wind go int32 Q16.16, and every within-group float bridge collapses,
so the WHOLE atmosphere/wave/wind/smoke/gas group is now cross-GPU deterministic
(only the downstream FIRE coupling remains a float bridge, S3).

Core (atmosphere_solver.cpp):
* Diffusion = Red-Black Gauss-Seidel in RESIDUAL/flux form (a quotient form has
  no fixed point under a truncating multiply at large mu*wsum). Per-cell
  Dinv = reciprocal_q16(quantize(1 + mu*wsum)), CACHED and rebuilt ONLY on changed
  cells (keyed on mu | masks | the 4 face permeabilities) — most ticks rebuild
  nothing. The GS increment is ROUND-TO-NEAREST (a truncating increment DC-sinks
  ~1.8%/transient; round-to-nearest brings it to ~float granularity).
* wind = -grad(atm+wave_p) is integer (the wave_p dequantize bridge collapsed);
  the *0.5 central diff is shr_round0 (signed).
* sponge/vacuum atmosphere scales -> mul_q16 (a deliberate sink, by design).

Bridge collapse (the headline — none remain in the S2 group):
1. wave->atmosphere transfer: a CONSERVATIVE INTEGER +/- PAIR (gather d once,
   atmosphere[i] += d, wave_p[i] -= d, the SAME round-to-nearest int) — exactly
   mass-neutral to the LSB, replacing the S2a one-sided mul_q16 deposit that
   DC-sank ~0.9%/200t.
2. wind read of wave_p -> integer.  3. smoke wind read -> integer (advection
   displacement + |wind|^2 diffusion).  4. n_smoke max_wind_sq -> integer max
   reduction.  5. W3/W5 water<->atmosphere -> int<->int (boil threshold compare,
   P*V mul_q16).  6. water head read of atmosphere -> integer-sourced bridge.
The ONE bridge S2 leaves open is FIRE: step_tail dequantizes atmosphere/wind to
float scratch, runs the float fire+temperature, re-quantizes the plume.

Render (the bug Erik flagged): pressure_overlay.py read `atmosphere + wave_p` as
RAW int32 counts (~65536x too large) -> saturated the top colour stop -> opaque
overlay painting over the smoke (wrong explosion colours AND invisible smoke).
FIX: dequantize BOTH to the pre-migration float scale (1.0..7.5 real, the colour
stops resolve as tuned). recorder.py dequantizes atmosphere/wind too.

Python: gamemap atmosphere/wind -> int32; new simulation/atmosphere_fixed.py
boundary helpers; field_edit "atmosphere" dtype (combine in real, store quantized);
combat O2 + gamemap find_burst_walls dequantize.

Gates (all green): P1 atmosphere+wind int32 bit-identical run-twice + cross-config
(field_ab_harness). P2 saturation (tests/test_atmosphere_saturation.py): sealed
box, strong blast, 3200 ticks -> atmosphere mass saturates then BIT-FROZEN to the
LSB; the revert-to-one-sided-mul_q16 RED check proves the +/- pair is what
conserves. GS residual bounded (~4e-4 peak) + converges in 8 sweeps like the float
GS. Feel vs float build: max|atm diff| 0.002, max|wind diff| 0.055 real units
(Q16.16 granularity). Full suite 357 green; both --auto exit 0 and render.

Branch only — atomic merge to main is Erik's tomorrow.

+++ cpp/src/smoke_dynamics.h
@@ -13,10 +13,12 @@
 // deterministic on every machine, so it is behaviour, not desync; smoke decay is
 // the tuning knob. NO flux form, NO limiter, NO outflow clamp.
 //
-// FLOAT BRIDGE (until S2c): the `wind` is still produced by the float atmosphere
-// (S2c not yet integer), so step() reads float wind_x/wind_y and quantizes the
-// wind*dt_adv DISPLACEMENT to Q16.16 at the boundary. After S2c the wind is
-// integer and this bridge closes (the displacement is computed integer-only).
+// S2c: the `wind` is now Q16.16 int32 (= -grad(atm+wave_p), computed by the
+// atmosphere solver) — the wind FLOAT BRIDGE is COLLAPSED. step() reads int32
+// wind_x/wind_y: the advection displacement -wind·dt_adv is an integer multiply
+// (dt_adv folded once in double, then quantized) and the wind-coupled diffusion
+// |wind|² is an integer square (Q.32) dequantized once per cell to the positive
+// d_eff·dt scalar coefficient. No per-cell float on the wind path.
 
 #include <cstdint>
 #include <vector>
@@ -42,8 +44,8 @@ public:
     // wind field from the atmosphere solver.
     void step(
         int32_t* smoke,            // S2b: int32 Q16.16 density (was float)
-        const float* wind_x,       // precomputed: -d/dx(atmosphere + wave_p) — FLOAT BRIDGE until S2c
-        const float* wind_y,       // precomputed: -d/dy(atmosphere + wave_p) — FLOAT BRIDGE until S2c
+        const int32_t* wind_x,     // S2c: Q16.16 int32 = -d/dx(atmosphere + wave_p)
+        const int32_t* wind_y,     // S2c: Q16.16 int32 = -d/dy(atmosphere + wave_p)
         const bool* obstacles,
         const bool* is_wall,
         const bool* is_vacuum,
